feat(clients/service): add CRUD methodes

+++ src/main/java/com/thifuge/kaala_plus/clients/ClientRepository.java
@@ -2,5 +2,5 @@ package com.thifuge.kaala_plus.clients;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientRepository extends JpaRepository<Client, Long> {
+public interface ClientRepository extends JpaRepository<Client, Integer> {
 }
